Update models and categories controller

+++ src/ProductCatalogService/Entities/Rating.cs
@@ -3,5 +3,5 @@
 public class Rating
 {
   public required string CustomerId { get; set; }
-  public int RatingValue { get; set; }
+  public double RatingValue { get; set; }
 }
